Notifications Permissions
I changed things around so that we ask to allow notifications when the
first med is added. I tried to do this changing as little existing code
as possible. If you like the very basic layout of the popup view, this
view could be used in place of alert views as well. But the pictures
and layout of the view should be changed; I don’t think it looks all
that good in its current state.